Broke up widgets and types from custom types
