+++ src/App.tsx
@@ -2,7 +2,7 @@ import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import React from "react";
-import { Alert, FlatList, Image, StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
+import { Alert, FlatList, Image, StyleSheet } from 'react-native';
 import 'react-native-gesture-handler';
 import { Provider } from 'react-redux';
 import EditProfile from "./1-Profile/Edit-Profile";
@@ -15,8 +15,8 @@ import { CircleList } from './2-Circles/CircleList';
 import { useAppDispatch, useAppSelector } from './TypesAndInterfaces/hooks';
 import store, { RootState } from './redux-store';
 import theme from './theme';
-import { CircleTabNavigator } from "./widgets";
-import { AppStackParamList, CIRCLE_DISPLAY_ROUTE_NAME, CIRCLE_LIST_ROUTE_NAME, CIRCLE_SEARCH_ROUTE_NAME, BOTTOM_TAB_NAVIGATOR_ROUTE_NAME, CIRCLE_NAVIGATOR_ROUTE_NAME, PRAYER_REQUEST_LIST_ROUTE_NAME, PRAYER_REQUEST_NAVIGATOR_ROUTE_NAME, PRAYER_REQUEST_DISPLAY_ROUTE_NAME } from './TypesAndInterfaces/custom-types';
+import { AppStackParamList, ROUTE_NAMES } from './TypesAndInterfaces/routes';
+import { CircleTabNavigator } from './2-Circles/circle-widgets';
 
 const Stack = createStackNavigator<AppStackParamList>();
 const Tab = createBottomTabNavigator();
@@ -29,9 +29,9 @@ const CircleTabNavigatorOptions = {
 const CircleStackNavigatorProp = () => {
   return (
     <Stack.Navigator screenOptions={{headerShown: false}}>
-      <Stack.Screen name={CIRCLE_LIST_ROUTE_NAME} component={CircleList} />
-      <Stack.Screen name={CIRCLE_DISPLAY_ROUTE_NAME} component={CircleDisplay} />
-      <Stack.Screen name={CIRCLE_SEARCH_ROUTE_NAME} component={CircleSearch} />
+      <Stack.Screen name={ROUTE_NAMES.CIRCLE_LIST_ROUTE_NAME} component={CircleList} />
+      <Stack.Screen name={ROUTE_NAMES.CIRCLE_DISPLAY_ROUTE_NAME} component={CircleDisplay} />
+      <Stack.Screen name={ROUTE_NAMES.CIRCLE_SEARCH_ROUTE_NAME} component={CircleSearch} />
     </Stack.Navigator>
   )
 }
@@ -42,7 +42,7 @@ const BottomTabNavigator = () => {
       <Tab.Navigator screenOptions={CircleTabNavigatorOptions}
           tabBar={props => <CircleTabNavigator {...props} />}
       >
-        <Tab.Screen name={CIRCLE_NAVIGATOR_ROUTE_NAME} component={CircleStackNavigatorProp} />
+        <Tab.Screen name={ROUTE_NAMES.CIRCLE_NAVIGATOR_ROUTE_NAME} component={CircleStackNavigatorProp} />
 
       </Tab.Navigator>
   );
@@ -57,7 +57,7 @@ const App = ():JSX.Element => {
             <Stack.Screen name="Login" component={Login} />
             <Stack.Screen name="Signup" component={Signup} />
             <Stack.Screen name="EditProfile" component={EditProfile} />
-            <Stack.Screen name={BOTTOM_TAB_NAVIGATOR_ROUTE_NAME} component={BottomTabNavigator} />
+            <Stack.Screen name={ROUTE_NAMES.BOTTOM_TAB_NAVIGATOR_ROUTE_NAME} component={BottomTabNavigator} />
           </Stack.Navigator>
         </NavigationContainer>
     </Provider>
